Fixed code scanning warnings

+++ examples/generate_hmac.c
@@ -94,7 +94,7 @@ int main(void) {
   assert(yrc == YHR_SUCCESS);
 
   printf("HMAC of data (%zu bytes) is:", hmac_data_len);
-  for (uint16_t i = 0; i < hmac_data_len; i++) {
+  for (size_t i = 0; i < hmac_data_len; i++) {
     printf(" %02x", hmac_data[i]);
   }
   printf("\n");
